refactor(backend): modularize migration commands & enhance readFile MCP tool with offset/size
